+++ flourish/__init__.py
@@ -164,6 +164,15 @@ class Flourish(object):
             path = self._paths[key]
             path.generate(report, tokens=[args])
 
+    def path_recipe(self, path):
+        handlers = self.get_handler_for_path(path)
+        for key, args in handlers:
+            path = self._paths[key]
+            recipe = path.get_recipe(tokens=[args])
+            if recipe:
+                return recipe
+        return None
+
     def set_global_context(self, global_context):
         self.global_context = global_context